refactor: final DDD cleanup and infrastructure isolation
- Delete deprecated legacy views, hooks, contexts, and root types
- Relocate config and bridge to dedicated infrastructure directory
- Strip api/client.ts to contain only Axios interceptor setup
- Update cross-module imports to target strictly defined domain boundaries
