+++ resource-app/frontend/components/Icons.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { RESOURCE_ICONS } from '../constants';
+import { RESOURCE_ICONS } from '../features/resource/constants';
 
 export const DynamicIcon = ({ name, className }: { name: string, className?: string }) => {
   const Icon = RESOURCE_ICONS[name] || RESOURCE_ICONS.DEFAULT;
