--- resource-app/frontend/context/AppContext.tsx
@@ -1,6 +1,0 @@
-// AppContext is now a thin shell kept for backward compatibility.
-// All domain state (resources, bookings, users, holidays)
-// now lives in their respective feature providers.
-// This file can be removed once all consumers are migrated.
-
-export {};
