+++ resource-app/frontend/features/user/views/AdminView.tsx
@@ -6,7 +6,8 @@ import { useResource } from '../../resource/context';
 import { Plus, Trash2, CheckCircle, Edit2, User, Shield } from 'lucide-react';
 import { cn } from '../../../utils/cn';
 import { Card, Button, Badge, PageLoader, EmptyState, Modal, Input, Label } from '../../../components/UI';
-import { BookingStatus, UserRole } from '../../../types';
+import { BookingStatus } from '../../booking/types';
+import { UserRole } from '../types';
 import { Resource } from '../../resource/types';
 import { format } from 'date-fns';
 import { CreateResourceView } from '../../resource/views/CreateResourceView';
