+++ resource-app/frontend/infrastructure/bridge.ts
