+++ resource-app/frontend/infrastructure/config.ts
