--- resource-app/frontend/services/geminiService.ts
@@ -1,2 +1,0 @@
-// This service is deprecated as all AI features have been removed.
-export {};
