--- resource-app/frontend/types.ts
@@ -1,8 +1,0 @@
-export type { ApiResponse } from './api/types';
-export { UserRole } from './features/user/types';
-export type { User } from './features/user/types';
-export type { PublicHoliday } from './features/holiday/types';
-export { ResourceType, RESOURCE_TYPES } from './features/resource/types';
-export type { Resource, FormField, ResourceUsageStats } from './features/resource/types';
-export { BookingStatus } from './features/booking/types';
-export type { Booking } from './features/booking/types';
