--- resource-app/frontend/views/SettingsView.tsx
@@ -1,2 +1,0 @@
-// File deleted. ProfileView is used instead.
-export {};
